+++ TabExpansionProfile.Invoke-Build.ps1
@@ -4,36 +4,39 @@
 	TabExpansion2 profile for Invoke-Build.
 
 .Description
-	Use this profile with the custom TabExpansion2
+	This script should be in the path. It is invoked on the first call of the
+	custom TabExpansion2. It adds code completers to the global option table.
 	https://farnet.googlecode.com/svn/trunk/PowerShellFar/TabExpansion2.ps1
 
-	Normally TabExpansion2.ps1 should be called in the very beginning of a
-	session from a PowerShell profile. This script should be placed to the
-	system path. It will be called on the first code completion.
-
-	This script adds completers for Task and File arguments of Build.ps1 and
+	This profile adds completers for Task and File arguments of Build.ps1 and
 	Invoke-Build.ps1
 #>
+
+# Common completer for Task of Build.ps1 and Invoke-Build.ps1.
+$completeTask = {
+	param($commandName, $parameterName, $wordToComplete, $commandAst, $boundParameters)
+
+	(& $commandName ?? -File ($boundParameters['File'])).Keys -like "$wordToComplete*" |
+	.{process{ New-CompletionResult $_ }}
+}
 
 # Common completer for File of Build.ps1 and Invoke-Build.ps1.
 $completeFile = {
 	param($commandName, $parameterName, $wordToComplete, $commandAst, $boundParameters)
-	Get-ChildItem -Directory -Name "$wordToComplete*"
+
+	Get-ChildItem -Directory -Name "$wordToComplete*" |
+	.{process{ New-CompletionResult $_ -ResultType ProviderContainer }}
+
 	if (!($boundParameters['Task'] -eq '**')) {
-		Get-ChildItem -File -Name "$wordToComplete*.ps1"
+		Get-ChildItem -File -Name "$wordToComplete*.ps1" |
+		.{process{ New-CompletionResult $_ -ResultType Command }}
 	}
 }
 
 # Add completers to the option table
 $TabExpansionOptions.CustomArgumentCompleters += @{
-	'Build.ps1:Task' = {
-		param($commandName, $parameterName, $wordToComplete, $commandAst, $boundParameters)
-		(Build.ps1 ?? -File ($boundParameters['File'])).Keys -like "$wordToComplete*"
-	}
-	'Invoke-Build.ps1:Task' = {
-		param($commandName, $parameterName, $wordToComplete, $commandAst, $boundParameters)
-		(Invoke-Build.ps1 ?? -File ($boundParameters['File'])).Keys -like "$wordToComplete*"
-	}
+	'Build.ps1:Task' = $completeTask
 	'Build.ps1:File' = $completeFile
+	'Invoke-Build.ps1:Task' = $completeTask
 	'Invoke-Build.ps1:File' = $completeFile
 }